+++ yellowstone-grpc-client-nodejs/src/index.ts
@@ -1,13 +1,4 @@
-/**
- * TypeScript/JavaScript client for gRPC Geyser.
- */
-
-import {
-  ChannelCredentials,
-  credentials,
-  ChannelOptions,
-  Metadata,
-} from "@grpc/grpc-js";
+/** TypeScript/JavaScript client for gRPC Geyser. */
 
 // Import generated gRPC client and types.
 import {
@@ -15,6 +6,8 @@ import {
   GetLatestBlockhashResponse,
   GeyserClient,
   IsBlockhashValidResponse,
+  SubscribeReplayInfoResponse,
+  SubscribeRequest,
   SubscribeRequestAccountsDataSlice,
   SubscribeRequestFilterAccounts,
   SubscribeRequestFilterBlocks,
@@ -23,28 +16,27 @@ import {
   SubscribeRequestFilterSlots,
   SubscribeRequestFilterTransactions,
   SubscribeUpdateTransactionInfo,
-  SubscribeReplayInfoResponse,
 } from "./grpc/geyser";
 
 // Reexport automatically generated types
 export {
   CommitmentLevel,
   SubscribeRequest,
-  SubscribeRequestAccountsDataSlice,
-  SubscribeRequestFilterAccounts,
-  SubscribeRequestFilterAccountsFilter,
-  SubscribeRequestFilterAccountsFilterMemcmp,
-  SubscribeRequestFilterAccountsFilterLamports,
-  SubscribeRequestFilterBlocks,
-  SubscribeRequestFilterBlocksMeta,
-  SubscribeRequestFilterEntry,
-  SubscribeRequestFilterSlots,
-  SubscribeRequestFilterTransactions,
   SubscribeRequest_AccountsEntry,
   SubscribeRequest_BlocksEntry,
   SubscribeRequest_BlocksMetaEntry,
   SubscribeRequest_SlotsEntry,
   SubscribeRequest_TransactionsEntry,
+  SubscribeRequestAccountsDataSlice,
+  SubscribeRequestFilterAccounts,
+  SubscribeRequestFilterAccountsFilter,
+  SubscribeRequestFilterAccountsFilterLamports,
+  SubscribeRequestFilterAccountsFilterMemcmp,
+  SubscribeRequestFilterBlocks,
+  SubscribeRequestFilterBlocksMeta,
+  SubscribeRequestFilterEntry,
+  SubscribeRequestFilterSlots,
+  SubscribeRequestFilterTransactions,
   SubscribeUpdate,
   SubscribeUpdateAccount,
   SubscribeUpdateAccountInfo,
@@ -55,6 +47,74 @@ export {
   SubscribeUpdateTransaction,
   SubscribeUpdateTransactionInfo,
 } from "./grpc/geyser";
+
+import { Duplex } from "stream";
+import * as napi from "./napi/index";
+
+export default class Client {
+  _insecureEndpoint: string;
+  _insecureXToken: string | undefined;
+  _channelOptions: napi.JsChannelOptions | undefined;
+
+  constructor(endpoint: string, xToken: string | undefined, channel_options: napi.JsChannelOptions | undefined) {
+    this._insecureEndpoint = endpoint;
+    this._insecureXToken = xToken;
+    this._channelOptions = channel_options;
+  }
+
+  async subscribe(): Promise<ClientDuplexStream> {
+    // Inner stream.Duplex config passed to both stream.Readable and Writable.
+    // See: https://nodejs.org/en/blog/feature/streams2#new-streamduplexoptions
+    const options = {
+      // Pass objects not bytes.
+      objectMode: true,
+      // Skip unnecessary buffer conversion for performance.
+      decodeStrings: false,
+      // TODO: Fine tune high watermark for performance and backpressure.
+      // highWaterMark: 16
+    };
+
+    return new Promise<ClientDuplexStream>((resolve, reject) => {
+      try {
+        resolve(new ClientDuplexStream(this._insecureEndpoint, this._insecureXToken, this._channelOptions, options));
+      } catch (err) {
+        reject(err)
+      }
+    });
+  }
+}
+
+class ClientDuplexStream extends Duplex {
+  _napiDuplexStream: napi.DuplexStream;
+
+  constructor(endpoint: string, xToken: string | undefined, channel_options: napi.JsChannelOptions | undefined, options: object | undefined) {
+    super({ ...options });
+    this._napiDuplexStream = new napi.DuplexStream(endpoint, xToken, channel_options);
+  }
+
+  async _read(_size: number) {
+    try {
+      const update = await this._napiDuplexStream.read();
+      this.push(update)
+    } catch (err) {
+      this.push(null); // Signal end of stream
+      this.destroy(err); // Handle resource cleanup
+    }
+  }
+
+  _write(chunk: object, _encoding: any, callback: any) {
+    try {
+      this._napiDuplexStream.write(chunk);
+      callback()
+    } catch (err) {
+      callback(err)
+    }
+  }
+}
+
+//--------------------------------------------------------------------------------
+// Use old WASM encoders for now.
+//--------------------------------------------------------------------------------
 
 // Import transaction encoding function created in Rust
 import * as wasm from "./encoding/yellowstone_grpc_solana_encoding_wasm";
@@ -90,221 +150,3 @@ export const txErrDecode = {
     return JSON.parse(wasm.decode_tx_error(buf));
   },
 };
-
-export default class Client {
-  _client: GeyserClient;
-  _insecureXToken: string | undefined;
-  constructor(
-    endpoint: string,
-    xToken: string | undefined,
-    channelOptions: ChannelOptions | undefined
-  ) {
-    let creds: ChannelCredentials;
-
-    const endpointURL = new URL(endpoint);
-    let port = endpointURL.port;
-    if (port == "") {
-      switch (endpointURL.protocol) {
-        case "https:":
-          port = "443";
-          break;
-        case "http:":
-          port = "80";
-          break;
-      }
-    }
-
-    // Check if we need to use TLS.
-    if (endpointURL.protocol === "https:") {
-      creds = credentials.combineChannelCredentials(
-        credentials.createSsl(),
-        credentials.createFromMetadataGenerator((_params, callback) => {
-          const metadata = new Metadata();
-          if (xToken !== undefined) {
-            metadata.add("x-token", xToken);
-          }
-          return callback(null, metadata);
-        })
-      );
-    } else {
-      creds = ChannelCredentials.createInsecure();
-      if (xToken !== undefined) {
-        this._insecureXToken = xToken;
-      }
-    }
-
-    this._client = new GeyserClient(
-      `${endpointURL.hostname}:${port}`,
-      creds,
-      channelOptions
-    );
-  }
-
-  private _getInsecureMetadata(): Metadata {
-    const metadata = new Metadata();
-    if (this._insecureXToken) {
-      metadata.add("x-token", this._insecureXToken);
-    }
-    return metadata;
-  }
-
-  async subscribe() {
-    return await this._client.subscribe(this._getInsecureMetadata());
-  }
-
-  async subscribeOnce(
-    accounts: { [key: string]: SubscribeRequestFilterAccounts },
-    slots: { [key: string]: SubscribeRequestFilterSlots },
-    transactions: { [key: string]: SubscribeRequestFilterTransactions },
-    transactionsStatus: { [key: string]: SubscribeRequestFilterTransactions },
-    entry: { [key: string]: SubscribeRequestFilterEntry },
-    blocks: { [key: string]: SubscribeRequestFilterBlocks },
-    blocksMeta: { [key: string]: SubscribeRequestFilterBlocksMeta },
-    commitment: CommitmentLevel | undefined,
-    accountsDataSlice: SubscribeRequestAccountsDataSlice[]
-  ) {
-    const stream = await this._client.subscribe(this._getInsecureMetadata());
-
-    await new Promise<void>((resolve, reject) => {
-      stream.write(
-        {
-          accounts,
-          slots,
-          transactions,
-          transactionsStatus,
-          entry,
-          blocks,
-          blocksMeta,
-          commitment,
-          accountsDataSlice,
-        },
-        (err: any) => {
-          if (err === null || err === undefined) {
-            resolve();
-          } else {
-            reject(err);
-          }
-        }
-      );
-    });
-
-    return stream;
-  }
-
-  async subscribeReplayInfo(): Promise<SubscribeReplayInfoResponse> {
-    return await new Promise<SubscribeReplayInfoResponse>((resolve, reject) => {
-      this._client.subscribeReplayInfo(
-        {},
-        this._getInsecureMetadata(),
-        (err, response) => {
-          if (err === null || err === undefined) {
-            resolve(response);
-          } else {
-            reject(err);
-          }
-        }
-      );
-    });
-  }
-
-  async ping(count: number): Promise<number> {
-    return await new Promise<number>((resolve, reject) => {
-      this._client.ping(
-        { count },
-        this._getInsecureMetadata(),
-        (err, response) => {
-          if (err === null || err === undefined) {
-            resolve(response.count);
-          } else {
-            reject(err);
-          }
-        }
-      );
-    });
-  }
-
-  async getLatestBlockhash(
-    commitment?: CommitmentLevel
-  ): Promise<GetLatestBlockhashResponse> {
-    return await new Promise<GetLatestBlockhashResponse>((resolve, reject) => {
-      this._client.getLatestBlockhash(
-        { commitment },
-        this._getInsecureMetadata(),
-        (err, response) => {
-          if (err === null || err === undefined) {
-            resolve(response);
-          } else {
-            reject(err);
-          }
-        }
-      );
-    });
-  }
-
-  async getBlockHeight(commitment?: CommitmentLevel): Promise<string> {
-    return await new Promise<string>((resolve, reject) => {
-      this._client.getBlockHeight(
-        { commitment },
-        this._getInsecureMetadata(),
-        (err, response) => {
-          if (err === null || err === undefined) {
-            resolve(response.blockHeight);
-          } else {
-            reject(err);
-          }
-        }
-      );
-    });
-  }
-
-  async getSlot(commitment?: CommitmentLevel): Promise<string> {
-    return await new Promise<string>((resolve, reject) => {
-      this._client.getSlot(
-        { commitment },
-        this._getInsecureMetadata(),
-        (err, response) => {
-          if (err === null || err === undefined) {
-            resolve(response.slot);
-          } else {
-            reject(err);
-          }
-        }
-      );
-    });
-  }
-
-  async isBlockhashValid(
-    blockhash: string,
-    commitment?: CommitmentLevel
-  ): Promise<IsBlockhashValidResponse> {
-    return await new Promise<IsBlockhashValidResponse>((resolve, reject) => {
-      this._client.isBlockhashValid(
-        { blockhash, commitment },
-        this._getInsecureMetadata(),
-        (err, response) => {
-          if (err === null || err === undefined) {
-            resolve(response);
-          } else {
-            reject(err);
-          }
-        }
-      );
-    });
-  }
-
-  async getVersion(): Promise<string> {
-    return await new Promise<string>((resolve, reject) => {
-      this._client.getVersion(
-        {},
-        this._getInsecureMetadata(),
-        (err, response) => {
-          if (err === null || err === undefined) {
-            resolve(response.version);
-          } else {
-            reject(err);
-          }
-        }
-      );
-    });
-  }
-}